fix(schema): use correct name for lookups
Without this change, the modified name of the type is sometimes used when
determining if a type should be included in the expanded set, which results in
odd behavior, or some types failing to be matched because their name has been
modified before the lookup call.

Here we ensure that the lookup of types is performed based on their name as
represented in the schema, rather than what we want to impose upon the name
based on the language we are working with.