chore: adjust ExpandTypes signature to allow for config changes
Without this change, each time the configuration changes, we need to update the
signature to allow for the change in scope of what the ExpandTypes method is
doing.  Here we modify that notion to allow the ExpandTypes to operate on
anything within the configuration.  This should mean that when the config
changes, we don't also need to modify each of the generator calls which utilize
this method.